[SIPM][SIFISC-18687] / 2 : Changement du type de donnée capital libéré
git-svn-id: https://svn.etat-de-vaud.ch/SVN_SOCLE/registre/unireg/trunk@120967 cb3efa0a-b8b0-9f43-a4c8-343d8b7e4362

+++ unireg/web/src/main/java/ch/vd/uniregctb/tiers/view/EntrepriseCivilView.java
@@ -1,7 +1,5 @@
 package ch.vd.uniregctb.tiers.view;
 
-import java.math.BigDecimal;
-
 import ch.vd.registre.base.date.RegDate;
 import ch.vd.uniregctb.tiers.MontantMonetaire;
 import ch.vd.uniregctb.type.FormeJuridiqueEntreprise;
@@ -16,7 +14,7 @@ public class EntrepriseCivilView {
 	private Integer numeroOfsSiege;
 	private String nomSiege;
 	private FormeJuridiqueEntreprise formeJuridique;
-	private BigDecimal capitalLibere;
+	private Long capitalLibere;
 	private String devise;
 	private String numeroIde;
 	private boolean inscriteRC;
@@ -82,11 +80,11 @@ public class EntrepriseCivilView {
 		this.formeJuridique = formeJuridique;
 	}
 
-	public BigDecimal getCapitalLibere() {
+	public Long getCapitalLibere() {
 		return capitalLibere;
 	}
 
-	public void setCapitalLibere(BigDecimal capitalLibere) {
+	public void setCapitalLibere(Long capitalLibere) {
 		this.capitalLibere = capitalLibere;
 	}
